changes added to dashboard

+++ Code/View/SavingsGUI.java
@@ -136,15 +136,14 @@ public class SavingsGUI extends JPanel {
 
         try {
         // Parse target and starting amount as doubles
-        Float targetAmount = Float.parseFloat(target);
-        Float startingAmountValue = Float.parseFloat(startingAmount);
+        Double targetAmount = Double.parseDouble(target);
+        Double startingAmountValue = Double.parseDouble(startingAmount);
         
         // Create a SavingsGoal object
         SavingsGoal newGoal = new SavingsGoal(name, targetAmount, deadline, startingAmountValue, notificationsEnabled);
 
             // Add the saving goal into the database table
             savingDao.addGoalIntoDatabase(newGoal);
-
             SavingsGoalDAO.setCurrentSavingsGoal(newGoal);
 
             // Show success message
